Rename one_dimension -> onedim_core
